Download bar initialization review. Performance check.
A bug was found on the initialization of the download bar prompting the
need to load it with the datatable for refreshing the data. A probable
bug due to the flash object persistance blocks the possibility to use
the download bar on the detailed view.
Leak remain on the googleLayer at least but doesn't impact the behaviour
of the browser.